requirements.txt
PyYAML
pandas
numpy
matplotlib
scikit-learn
requests
beautifulsoup4
black
geopandas
